Validate against special characters in Message
If a message contains special characters it may be badly encoded when
sent over SMS (depending on the carrier and device).

There are some characters which are frequently bad, easily replaceable
with ascii characters and are hard to spot by eye.

This commit adds a validator to prevent them from being saved.